Removed hubot as a dependency
This doesn't appear to be required and there would really be no way as far as I know to run this hubot script without hubot itself and is giving warnings when installing because we are on hubot v3 already.  It doesn't appear other scripts list hubot as a dependency so rather than do some sort of 3.0 >=2.x style thing, I figured it would be best to remove the dependency altogether.